Small Shopping list application
This is another practice application i have built, there is really minimal features and design to the application as it serves as practice for me and nothing more. These smaller projects are not worthy of inclusion on my portfolio but they are great for making sure i am learning and able to implement concepts that i am learning. There is nothing new in this app to me previous few this is just more practice for things such as lifting up state, making use of the children prop for component reusablilty, controlled elements for inputs and derived state for my listEls array as i am now much more clear on the use of derived state being that if a state changes the derived state will be automatically updated when the component re-renders and therefor update the derived state.